chore(tests): move comment action tests in their own file
