+++ mergify_engine/tests/functional/test_engine.py
@@ -202,88 +202,6 @@ class TestEngineV2Scenario(base.FunctionalTestBase):
         self.assertEqual("CHANGES_REQUESTED", comments[-1].state)
         self.assertEqual("WTF?", comments[-1].body)
 
-    def test_comment(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "comment",
-                    "conditions": [f"base={self.master_branch_name}"],
-                    "actions": {"comment": {"message": "WTF?"}},
-                }
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p, _ = self.create_pr()
-
-        p.update()
-        comments = list(p.get_issue_comments())
-        self.assertEqual("WTF?", comments[-1].body)
-
-        # Add a label to trigger mergify
-        self.add_label(p, "stable")
-
-        # Ensure nothing changed
-        new_comments = list(p.get_issue_comments())
-        self.assertEqual(len(comments), len(new_comments))
-        self.assertEqual("WTF?", new_comments[-1].body)
-
-        # Add new commit to ensure Summary get copied and comment not reposted
-        open(self.git.tmp + "/new_file", "wb").close()
-        self.git("add", self.git.tmp + "/new_file")
-        self.git("commit", "--no-edit", "-m", "new commit")
-        self.git(
-            "push",
-            "--quiet",
-            "fork",
-            self.get_full_branch_name("fork/pr%d" % self.pr_counter),
-        )
-
-        self.wait_for("pull_request", {"action": "synchronize"})
-
-        # Ensure nothing changed
-        new_comments = list(p.get_issue_comments())
-        self.assertEqual(len(comments), len(new_comments))
-        self.assertEqual("WTF?", new_comments[-1].body)
-
-    def test_comment_backwardcompat(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "comment",
-                    "conditions": [f"base={self.master_branch_name}"],
-                    "actions": {"comment": {"message": "WTF?"}},
-                }
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p, _ = self.create_pr()
-
-        p.update()
-        comments = list(p.get_issue_comments())
-        self.assertEqual("WTF?", comments[-1].body)
-
-        # Override Summary with the old format
-        pull = context.Context(self.cli_integration, p.raw_data, {})
-        check_api.set_check_run(
-            pull,
-            "Summary",
-            "completed",
-            "success",
-            output={"title": "whatever", "summary": "erased"},
-        )
-
-        # Add a label to trigger mergify
-        self.add_label(p, "stable")
-
-        # Ensure nothing changed
-        new_comments = list(p.get_issue_comments())
-        self.assertEqual(len(comments), len(new_comments))
-        self.assertEqual("WTF?", new_comments[-1].body)
-
     def test_close(self):
         rules = {
             "pull_request_rules": [
